requirements.txt
tensorflow==2.10.1
keras==2.10.0
numpy==1.23.5
Pillow>=9.5.0
ffmpeg-python==0.2.0
tqdm>=4.66.0
matplotlib>=3.5.0
